requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
web3==7.14.1
python-dotenv==1.0.0
py-solc-x==2.0.2
